+++ users/users.mutations.js
@@ -1,5 +1,6 @@
 import client from "../client";
 import bcrypt from "bcrypt";
+import jwt from "jsonwebtoken";
 
 export default {
   Mutation: {
@@ -29,6 +30,38 @@ export default {
       } catch (e) {
         return e;
       }
+    },
+    login: async (_, { username, password }) => {
+      // find suer with args.username
+      const user = await client.user.findFirst({
+        where: {
+          username,
+        },
+      });
+      if (!user) {
+        return {
+          ok: false,
+          error: "User not found.",
+        };
+      }
+      // check password with arg.password
+      const passwordOk = await bcrypt.compare(password, user.password);
+      if (!passwordOk) {
+        return {
+          ok: false,
+          error: "Incorrect password.",
+        };
+      }
+      // issued a token and send it to the user
+      const token = await jwt.sign(
+        { id: user.id },
+        process.env.JWT_PRIVATE_KEY
+      );
+
+      return {
+        ok: true,
+        token,
+      };
     },
   },
 };